cmdline: move cmdline to x86
Now risc-v doesn't support cmdline

Tracked-On: #8805
